fix: change wrapper class to primitive type (#196)
* fix: change wrapper class to primitive type

* fix: model entity id to wrapper class

+++ mypage-service/src/main/java/com/yapp/crew/dto/request/UserReportRequestDto.java
@@ -9,10 +9,10 @@ import lombok.NoArgsConstructor;
 public class UserReportRequestDto {
 
 	@NotNull
-	private Long userId;
+	private long userId;
 
 	@NotNull
-	private Long reportType;
+	private int reportType;
 
 	private String content;
